Consolidate functions for components. Add buttons for danger, info and option.
